requirements.txt
sqlalchemy < 1.4.0
aiohttp==3.7.3
pytest==6.2.2
SQLAlchemy-Utils==0.36.8
python-decouple==3.4
mysqlclient==2.0.3
pytest_mock==3.5.1
PyJWT==2.0.1
pytest-cov==2.11.1
pytest-asyncio==0.14.0